backend/requirements.txt
django==3.2.4
djangorestframework==3.12.4
drf-yasg==1.20.0
psycopg2==2.8.6
django-filter==2.4.0
django-q==1.3.7
redis==3.5.3
django-cors-headers==3.7.0
